requirements.txt
Django==5.2
weasyprint==68.1
Pillow==12.2.0
django-storages==1.14.4
boto3==1.38.0
python-decouple==3.8
whitenoise==6.9.0
gunicorn==23.0.0
django-axes==7.0.1
cryptography==44.0.2
mammoth==1.9.0
anthropic
